Removed old CompilationResolver
