--- src/Dotnet.Script.DependencyModel/Compilation/CompilationDependencyResolver2.cs
@@ -1,28 +1,0 @@
-using System.Collections.Generic;
-using Dotnet.Script.DependencyModel.Logging;
-using Dotnet.Script.DependencyModel.ProjectSystem;
-
-namespace Dotnet.Script.DependencyModel.Compilation
-{
-    public class CompilationDependencyResolver2
-    {
-        private readonly ScriptProjectProvider _scriptProjectProvider;
-        private readonly ICompilationReferenceReader _compilationReferenceReader;
-
-        public CompilationDependencyResolver2(LogFactory logFactory) : this(new ScriptProjectProvider(logFactory), new CompilationReferencesReader(logFactory), logFactory)
-        {
-        }
-
-        public CompilationDependencyResolver2(ScriptProjectProvider scriptProjectProvider, ICompilationReferenceReader compilationReferenceReader, LogFactory logFactory)
-        {
-            _scriptProjectProvider = scriptProjectProvider;
-            _compilationReferenceReader = compilationReferenceReader;
-        }
-
-        public IEnumerable<CompilationReference> GetDependencies(string targetDirectory, IEnumerable<string> scriptFiles, bool enableScriptNugetReferences, string defaultTargetFramework = "net46")
-        {
-            var projectFileInfo = _scriptProjectProvider.CreateProject(targetDirectory, scriptFiles, defaultTargetFramework, enableScriptNugetReferences);
-            return _compilationReferenceReader.Read(projectFileInfo);
-        }
-    }
-}
