+++ src/Dotnet.Script.DependencyModel/Compilation/CompilationReference.cs
@@ -1,0 +1,12 @@
+namespace Dotnet.Script.DependencyModel.Compilation
+{
+    public class CompilationReference
+    {
+        public CompilationReference(string path)
+        {
+            Path = path;
+        }
+
+        public string Path { get; }
+    }
+}
